menu_card_7C028.c up to PS1_SetNoCard

+++ include/card.h
@@ -72,11 +72,11 @@ void LoadGameOnDisk(u8 slot);
 s32 LoadInfoGame(u8 slot);
 void FUN_8016bbe4(void);
 u8 PS1_GetNbreSave3(u8 param_1);
-s16 PS1_CardFilenameChecksumChanged(void);
-u8 PS1_TestCardZero(void);
+s32 PS1_CardFilenameChecksumChanged(void);
+s32 PS1_TestCardZero(void);
 u8 PS1_CardUnformatted(void);
 u8 PS1_GetNbreSave2(void);
-s16 FUN_8016be9c(void);
+s32 FUN_8016be9c(void);
 void FUN_8016bec0(void);
 
 #endif
